\name{flowCL_archive}
\alias{flowCL_archive}
\docType{data}
\title{
    A list of data used to build a local archive.
}
\description{
    A list of data used to build a local archive.
}
\usage{
    data(flowCL_archive)
}
\value{
    Loads the a pre-built archive.
}
\author{
    Justin Meskas
}
\keyword{
    datasets
}
