\name{test.flowCL.connection}
\alias{test.flowCL.connection}

\title{
    A function to check that a connection exists between the user and the ontology server.
}

\description{
    A function to check that a connection exists between the user and the ontology server.
}

\usage{
    test.flowCL.connection()
}

\author{ 
    Justin Meskas
}

\examples{
    test.flowCL.connection()
}


